# api/docs/swagger.yaml
definitions:
  github_com_coolbit-in_alone_openai.ChatCompletionMessage:
    properties:
      completion_tokens:
        type: integer
      content:
        type: string
      conversation_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      prompt_tokens:
        type: integer
      role:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  openai.Conversation:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/github_com_coolbit-in_alone_openai.ChatCompletionMessage'
        type: array
      name:
        type: string
      system_role_id:
        type: integer
      updatedAt:
        type: string
    type: object
  openai.SystemRole:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  title: Phantom Horse API
  version: v1.0
paths:
  /conversations:
    get:
      consumes:
      - application/json
      description: List conversations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/openai.Conversation'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List conversations
      tags:
      - conversation
    post:
      consumes:
      - application/json
      description: Add conversation
      parameters:
      - description: Conversation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/openai.Conversation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add conversation
      tags:
      - conversation
  /conversations/{conversation_id}:
    get:
      consumes:
      - application/json
      description: Get conversation
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.Conversation'
        "400":
          description: Bad Request
          schema: {}
      summary: Get conversation
      tags:
      - conversation
  /messages:
    post:
      consumes:
      - application/json
      description: Add messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add messages
      tags:
      - message
  /messages/{id}:
    get:
      consumes:
      - application/json
      description: Get message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_coolbit-in_alone_openai.ChatCompletionMessage'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get message
      tags:
      - message
  /system_roles:
    get:
      consumes:
      - application/json
      description: List system roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/openai.SystemRole'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List system roles
      tags:
      - system_role
    post:
      consumes:
      - application/json
      description: Add system role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.SystemRole'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add system role
      tags:
      - system_role
  /system_roles/{id}:
    get:
      consumes:
      - application/json
      description: Get system role
      parameters:
      - description: System Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.SystemRole'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get system role
      tags:
      - system_role
swagger: "2.0"